[Order] Use wait_asyncio_next_cycle
