+++ octobot_trading/orders/types/trailing/trailing_stop_order.py
@@ -17,7 +17,7 @@ import asyncio
 from asyncio import wait_for
 
 from octobot_trading.data.order import Order
-from octobot_trading.enums import TradeOrderSide, ExchangeConstantsMarketPropertyColumns
+from octobot_trading.enums import TradeOrderSide
 
 
 class TrailingStopOrder(Order):
