+++ requirements.txt
@@ -5,7 +5,7 @@ numpy==1.18.4
 # Drakkar-Software requirements
 OctoBot-Backtesting>=1.5.14, <1.6
 OctoBot-Channels>=1.4.8, <1.5
-OctoBot-Commons>=1.3.34, <1.4
+OctoBot-Commons>=1.3.36, <1.4
 OctoBot-Tentacles-Manager>=2.1.5, <2.2
 
 # Exchange connection requirements
