+++ tests/orders/types/__init__.py
@@ -13,6 +13,3 @@
 #
 #  You should have received a copy of the GNU Lesser General Public
 #  License along with this library.
-
-async def ensure_filled():
-    pass
