+++ tests/orders/types/limit/test_buy_limit_order.py
@@ -17,11 +17,12 @@ import asyncio
 
 import pytest
 
-from octobot_trading.enums import TradeOrderType, OrderStatus
+from octobot_trading.enums import TradeOrderType
+from octobot_commons.asyncio_tools import wait_asyncio_next_cycle
 from tests import event_loop
 from tests.exchanges import simulated_trader, simulated_exchange_manager
 from tests.orders import buy_limit_order
-from tests.orders.types import ensure_filled
+
 from tests.util.random_numbers import random_price, random_quantity, random_recent_trade, random_timestamp
 
 pytestmark = pytest.mark.asyncio
@@ -44,15 +45,15 @@ async def test_buy_limit_order_trigger(buy_limit_order):
     price_events_manager.handle_recent_trades(
         [random_recent_trade(price=random_price(min_value=order_price + 1),
                              timestamp=buy_limit_order.timestamp)])
-    await asyncio.create_task(ensure_filled())
+    await wait_asyncio_next_cycle()
     assert not buy_limit_order.is_filled()
     price_events_manager.handle_recent_trades(
         [random_recent_trade(price=order_price,
                              timestamp=buy_limit_order.timestamp - 1)])
-    await asyncio.create_task(ensure_filled())
+    await wait_asyncio_next_cycle()
     assert not buy_limit_order.is_filled()
     price_events_manager.handle_recent_trades([random_recent_trade(price=order_price,
                                                                    timestamp=buy_limit_order.timestamp)])
 
-    await asyncio.create_task(ensure_filled())
+    await wait_asyncio_next_cycle()
     assert buy_limit_order.is_filled()
