+++ tests/orders/types/market/test_sell_market_order.py
@@ -18,11 +18,12 @@ import asyncio
 
 import pytest
 
+from octobot_commons.asyncio_tools import wait_asyncio_next_cycle
 from octobot_trading.enums import TradeOrderType, OrderStatus
 from tests import event_loop
 from tests.exchanges import simulated_trader, simulated_exchange_manager
 from tests.orders import sell_market_order
-from tests.orders.types import ensure_filled
+
 from tests.util.random_numbers import random_price, random_quantity, random_recent_trade, random_timestamp
 
 pytestmark = pytest.mark.asyncio
@@ -45,5 +46,5 @@ async def test_sell_market_order_trigger(sell_market_order):
     price_events_manager.handle_recent_trades(
         [random_recent_trade(price=random_price(),
                              timestamp=sell_market_order.timestamp)])
-    await asyncio.create_task(ensure_filled())
+    await wait_asyncio_next_cycle()
     assert sell_market_order.is_filled()
